Refactor XsdValidator to XmlSchemesValidator
