--- Resources/Public/Xsd/test.xsd
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" version="1.1" xmlns="https://www.homepage-webhilfe.de/XML/XSD/" targetNamespace="https://www.homepage-webhilfe.de/XML/XSD/" elementFormDefault="qualified">
-     
-</xs:schema>
-
-
-
